🐛 Move validations constants to single folder
See #752
