+++ .changeset/spotty-doors-think.md
@@ -1,0 +1,5 @@
+---
+'myst-frontmatter': patch
+---
+
+Change the validators to avoid circular dependencise in CJS
