+++ packages/myst-frontmatter/src/index.ts
@@ -1,3 +1,4 @@
+export * from './constants.js';
 export * from './affiliations/index.js';
 export * from './biblio/index.js';
 export * from './contributors/index.js';
